pyproject.toml
[project]
name = "Entrega_RAG"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "redis>=5.0.0",
    "huggingface>=0.0.1",
    "datasets>=3.2.0",
    "google-cloud-aiplatform>=1.79.0",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.4.1",
    "tenacity>=9.0.0",
    "db-dtypes>=1.4.1",
]


